+++ scripts/__init__.py
@@ -6,7 +6,6 @@ from .extraction import (
     parse,
     namespaces,
 )
-from .utils import read_filter_terms
 
 __all__ = [
     "schemas",
@@ -15,5 +14,4 @@ __all__ = [
     "load_alias_entity_prior_probabilities",
     "parse",
     "namespaces",
-    "read_filter_terms",
 ]
